tmp/vm-filesystems/ubuntu-fresh/home/hacker/Projects/VMKit/vmkit/migration.py
#!/usr/bin/env python3
"""
VM migration support

Provides live migration capabilities for VMs between hosts.
"""

import libvirt
import subprocess
from typing import Dict, List, Optional
import logging

from .core import VMError

logger = logging.getLogger(__name__)


class MigrationManager:
    """
    Manages VM migration between hosts
    """
    
    def __init__(self):
        self._conn = None
    
    @property
    def conn(self) -> libvirt.virConnect:
        """Lazy libvirt connection"""
        if self._conn is None:
            try:
                self._conn = libvirt.open("qemu:///system")
            except libvirt.libvirtError as e:
                raise VMError(f"Failed to connect to libvirt: {e}")
        return self._conn
    
    def migrate_vm(self, vm_name: str, destination_uri: str, 
                   live: bool = True, persistent: bool = True) -> None:
        """
        Migrate a VM to another host
        
        Args:
            vm_name: Name of VM to migrate
            destination_uri: Libvirt URI of destination host
            live: Perform live migration
            persistent: Make migration persistent on destination
        """
        # Placeholder implementation
        raise NotImplementedError("Migration support coming soon")
    
    def check_migration_compatibility(self, vm_name: str, destination_uri: str) -> Dict:
        """Check if VM can be migrated to destination"""
        # Placeholder implementation
        raise NotImplementedError("Migration compatibility checking coming soon")
