tmp/vm-filesystems/ubuntu-fresh/home/hacker/Projects/VMKit/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools-scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "vmkit"
version = "0.1.0"
description = "Simple Python abstraction for libvirt/KVM with Secure Boot support"
readme = "README.md"
authors = [
    {name = "VMKit", email = "vmkit@example.com"},
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators", 
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9", 
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: System :: Systems Administration",
    "Topic :: System :: Emulators",
]
keywords = ["libvirt", "kvm", "qemu", "virtualization", "secure-boot", "uefi"]
requires-python = ">=3.8"
dependencies = [
    "libvirt-python>=9.0.0",
    "click>=8.0.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.910",
]
iso = [
    "pycdlib>=1.13.0",
]
all = [
    "pycdlib>=1.13.0",
]

[project.urls]
"Homepage" = "https://github.com/example/vmkit"
"Bug Reports" = "https://github.com/example/vmkit/issues"
"Source" = "https://github.com/example/vmkit"

[project.scripts]
vmkit = "vmkit.cli:vmkit"

[tool.setuptools]
packages = ["vmkit"]

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
